+++ src/components/ColorBarLegend.tsx
@@ -1,4 +1,5 @@
 import * as React from 'react';
+
 import makeStyles from "@material-ui/core/styles/makeStyles";
 
 const useStyles = makeStyles(theme => ({
@@ -29,6 +30,26 @@ const useStyles = makeStyles(theme => ({
     }
 ));
 
+const LABEL_CONTAINER_STYLE = {
+    fontSize: 'x-small',
+    fontWeight: 'bold',
+    width: '100%',
+    display: 'flex',
+    flexWrap: 'nowrap',
+} as React.CSSProperties;
+
+const TITLE_CONTAINER_STYLE = {
+    ...LABEL_CONTAINER_STYLE,
+    justifyContent: 'center',
+    paddingBottom: '0.5em',
+} as React.CSSProperties;
+
+const TICKS_CONTAINER_STYLE = {
+    ...LABEL_CONTAINER_STYLE,
+    justifyContent: 'space-between',
+} as React.CSSProperties;
+
+const LEGEND_CONTAINER_STYLE: React.CSSProperties = {padding: '0.5em', color: 'black'};
 
 interface ColorBarLegendProps {
     name: string;
@@ -38,6 +59,7 @@ interface ColorBarLegendProps {
     maxValue: number;
     width?: number | string;
     height?: number | string;
+    numTicks?: number;
 }
 
 export default function ColorBarLegend({name, units, imageData, minValue, maxValue, width, height}: ColorBarLegendProps) {
@@ -51,9 +73,25 @@ export default function ColorBarLegend({name, units, imageData, minValue, maxVal
         console.log("Label clicked!");
     };
 
+
+    const {name, units, imageData, minValue, maxValue} = this.props;
+    let {width, height, numTicks} = this.props;
     width = width || '240px';
     height = height || '24px';
-    const midValue = (minValue + maxValue) / 2;
+    // TODO: move logic out of here
+    numTicks = numTicks || 5;
+    const fractionDigits = Math.min(10, Math.max(2, Math.round(-Math.log10(maxValue - minValue))));
+    const delta = (maxValue - minValue) / (numTicks - 1);
+    const ticks = new Array(numTicks);
+    for (let i = 0; i < numTicks; i++) {
+        let v = minValue + i * delta;
+        let vr = Math.round(v);
+        if (Math.abs(vr - v) < 1e-8) {
+            ticks[i] = (<span key={i}>{vr}</span>);
+        } else {
+            ticks[i] = (<span key={i}>{v.toFixed(fractionDigits)}</span>);
+        }
+    }
     return (
         <div className={'ol-control ' + classes.container}>
             <div className={classes.title}>
@@ -63,13 +101,14 @@ export default function ColorBarLegend({name, units, imageData, minValue, maxVal
                 src={`data:image/png;base64,${imageData}`}
                 width={width}
                 height={height}
-                alt={'Legend'}
                 onClick={handleImageClick}
             />
-            <div className={classes.label} onClick={handleLabelClick}>
-                <span>{+minValue.toFixed(2)}</span>
-                <span>{+midValue.toFixed(2)}</span>
-                <span>{+maxValue.toFixed(2)}</span>
+            <div
+                className={classes.label}
+                style={TICKS_CONTAINER_STYLE}
+                onClick={handleLabelClick}
+            >
+                {ticks}
             </div>
         </div>
     );