+++ src/components/ol/types.ts
@@ -1,0 +1,45 @@
+export { default as OlBaseObject } from 'ol/Object';
+export { default as OlMap } from 'ol/Map';
+export { MapOptions as OlMapOptions } from 'ol/PluggableMap';
+export { default as OlMapBrowserEvent } from 'ol/MapBrowserEvent';
+export { default as OlView } from 'ol/View';
+export { ViewOptions as OlViewOptions } from 'ol/View';
+export { Control as OlControl } from 'ol/control';
+export { GeoJSON as OlGeoJSONFormat } from 'ol/format';
+export { default as OlTileLayer } from 'ol/layer/Tile';
+export { Options as OlTileLayerOptions } from 'ol/layer/Tile';
+export { default as OlTileGrid } from 'ol/tilegrid/TileGrid';
+export { Options as OlTileGridOptions } from 'ol/tilegrid/TileGrid';
+export { default as OlVectorLayer } from 'ol/layer/Vector';
+export { Options as OlVectorLayerOptions } from 'ol/layer/Vector';
+export { default as OlVectorSource } from 'ol/source/Vector';
+export { default as OlXYZSource } from 'ol/source/XYZ';
+export { Geometry as OlGeometry } from 'ol/geom';
+export { default as OlSimpleGeometry } from 'ol/geom/SimpleGeometry';
+export { default as OlGeometryType } from 'ol/geom/GeometryType';
+export { default as OlPolygonGeometry, fromCircle as olPolygonFromCircle } from 'ol/geom/Polygon';
+export { default as OlCircleGeometry } from 'ol/geom/Circle';
+export { Extent as OlExtent } from 'ol/extent';
+export { default as OlDrawInteraction } from 'ol/interaction/Draw';
+export { Options as OlDrawInteractionOptions } from 'ol/interaction/Draw';
+export { DrawEvent as OlDrawEvent } from 'ol/interaction/Draw';
+export { default as OlSelectInteraction } from 'ol/interaction/Select';
+export { Options as OlSelectInteractionOptions } from 'ol/interaction/Select';
+export { SelectEvent as OlSelectEvent } from 'ol/interaction/Select';
+export { default as OlFeature } from 'ol/Feature';
+export { default as OlRenderFeature } from 'ol/render/Feature';
+export { Color as OlColor } from 'ol/color';
+export { default as OlStyle } from 'ol/style/Style';
+export { default as OlFillStyle } from 'ol/style/Fill';
+export { default as OlStrokeStyle } from 'ol/style/Stroke';
+export { default as OlCircleStyle } from 'ol/style/Circle';
+export { default as OlOSMSource } from 'ol/source/OSM';
+export { default as OlEvent } from 'ol/events/Event';
+export { default as OlRenderEvent } from 'ol/render/Event';
+export { default as OlScaleLineControl } from 'ol/control/ScaleLine';
+export { Options as OlScaleLineControlOptions } from 'ol/control/ScaleLine';
+export {
+    get as olProjGet,
+    transformExtent as olProjTransformExtent,
+    fromLonLat as olProjFromLonLat
+} from 'ol/proj'