+++ src/reducers/dataReducer.ts
@@ -1,4 +1,5 @@
-import * as ol from 'openlayers';
+import {  OlVectorLayer } from '../components/ol/types';
+
 
 import { DataState, newDataState } from '../states/dataState';
 import { storeUserServers } from '../states/userSettings';
@@ -9,7 +10,7 @@ import {
     REMOVE_TIME_SERIES_GROUP,
     UPDATE_COLOR_BARS,
     UPDATE_DATASETS,
-    UPDATE_TIME_SERIES, UPDATE_DATASET_PLACE_GROUP, REMOVE_USER_PLACE, REMOVE_ALL_USER_PLACES
+    UPDATE_TIME_SERIES, UPDATE_DATASET_PLACE_GROUP, REMOVE_USER_PLACE, REMOVE_ALL_USER_PLACES, UPDATE_SERVER_INFO
 } from '../actions/dataActions';
 import { MAP_OBJECTS } from '../states/controlState';
 import { newId } from '../util/id';
@@ -22,6 +23,9 @@ export function dataReducer(state: DataState, action: DataAction): DataState {
         state = newDataState();
     }
     switch (action.type) {
+        case UPDATE_SERVER_INFO: {
+            return {...state, serverInfo: action.serverInfo};
+        }
         case UPDATE_DATASETS: {
             return {...state, datasets: action.datasets};
         }
@@ -135,7 +139,7 @@ export function dataReducer(state: DataState, action: DataAction): DataState {
 
 function removeUserPlacesFromLayer(userPlaceIds: string[]) {
     if (MAP_OBJECTS.userLayer) {
-        const userLayer = MAP_OBJECTS.userLayer as ol.layer.Vector;
+        const userLayer = MAP_OBJECTS.userLayer as OlVectorLayer;
         const source = userLayer.getSource();
         userPlaceIds.forEach(placeId => {
             const feature = source.getFeatureById(placeId);